pyproject.toml
[project]
name = "hands_on_airflow_on_cloud_composer"
version = "0.1.0"
description = "Airflow DAGs for Cloud Composer hands-on example"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.8"
dependencies = [
    "apache-airflow>=2.5.0", # Adjust version as needed for your Composer environment
    "apache-airflow-providers-google>=8.0.0", # Adjust version as needed
    "pendulum>=2.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]

[tool.black]
line-length = 79

[tool.pytest.ini_options]
pythonpath = [
  "."
]
testpaths = [
    "tests",
]
